db: replace processBatch() with db.CopyFromLines()
Add a new helper function for the COPY FROM logic, to make it easier to
test. The new tests are an attempt to capture the current behavior, but
note that the current behavior may not necessarily be correct/desirable.

To run the new tests, `go test` must be provided with a conninfo string
pointing to the database under test; see the README.