db: correctly count multiline rows
As mentioned in issues #19 and #50, our COPY FROM implementation assumes
that one line of CSV corresponds to one row, but that's not true -- a
quoted string may be spread over multiple lines. Fix our reported row
count by looking at the result of the COPY operation.

This does NOT solve the more general issue of multiline rows, which is
that if the batch boundary comes down in the middle of a row, we'll
fail. But it is a step towards more correct behavior.