batch: move to bufio.Reader and net.Buffers
Unlike the prior lib/pq implementation, PgConn.CopyFrom() does not
perform any internal buffering to reduce the number of network writes,
so the current implementation leads to one-write-per-line and a
significant slowdown. The overhead of Fprintln()ing to an io.Pipe is not
helping things, either.

Instead, read the accumulated lines directly into a net.Buffers
instance. Its Read() implementation interacts well with CopyFrom() --
each network write will push as much data as is available. Since the
data needs to include the end-of-line terminators, switch from
bufio.Scanner to the lower-level bufio.Reader API, which won't strip our
line endings. This is a more verbose interface than Scanner, but it
gives us close to full control over how and when copies are made, for an
even bigger performance improvement. db.CopyFromLines() now takes an
io.Reader directly (and the io.Pipe has disappeared completely).

There is an additional compatibility break introduced here, which is
that --token-size has been removed (it's no longer applicable to the
implementation now that bufio.Scanner is gone).