batch.Scan: don't split multi-line rows across batches
As mentioned in 8da678ea, the batching algorithm assumes that one row
consists of exactly one line. If a row contains multi-line quoted column
values, and we happen to split that row across multiple batches, we'll
fail the COPY.

Add a naive CSV parser to Scan() which prevents this accidental
splitting. This parser searches the incoming CSV for the QUOTE and
ESCAPE characters (which can be customized by the new -quote and -escape
flags to the tool), and prevents Scan() from sending out a new batch if
it detects that we're still inside a quoted value.

Because it's imperative that our parser matches the Postgres CSV parser
exactly (and the rules for parsing are not as intuitive as the rules for
production), a wide variety of test cases have been added. Running the
suite with TEST_CONNINFO will additionally run these test cases against
a live Postgres server, as a sanity check to ensure that each case was
coded correctly.

Fixes #19 and #50.

This implementation adds a significant CPU hit, in the form of naive
iteration over every byte in the input. I duplicated each benchmark, to
measure the case where QUOTE and ESCAPE are the same and the case where
they are different. (The code paths diverge enough that it may be useful
to optimize them separately.)

For the following machine (with CPU scaling disabled to the best of my
knowledge):

    goos: linux
    goarch: amd64
    pkg: github.com/timescale/timescaledb-parallel-copy/internal/batch
    cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz

benchstat reports the following changes as of this commit, when compared
to the previous one:

    $ benchstat original.txt multiline.txt
    name                                                    old time/op  new time/op  delta
    Scan/warmup_(disregard)_(standard_escapes)-8             294µs ±23%   623µs ± 8%  +111.59%  (p=0.008 n=5+5)
    Scan/warmup_(disregard)_(custom_escapes)-8               294µs ±23%   617µs ± 1%  +109.60%  (p=0.008 n=5+5)
    Scan/no_quotes_(standard_escapes)-8                      273µs ± 2%   625µs ± 2%  +129.08%  (p=0.008 n=5+5)
    Scan/no_quotes_(custom_escapes)-8                        273µs ± 2%   633µs ± 1%  +132.08%  (p=0.008 n=5+5)
    Scan/some_quotes_at_the_beginning_(standard_escapes)-8   281µs ± 6%   835µs ± 2%  +196.96%  (p=0.008 n=5+5)
    Scan/some_quotes_at_the_beginning_(custom_escapes)-8     281µs ± 6%   850µs ± 2%  +202.22%  (p=0.008 n=5+5)
    Scan/some_quotes_in_the_middle_(standard_escapes)-8      280µs ± 5%   856µs ± 2%  +205.66%  (p=0.008 n=5+5)
    Scan/some_quotes_in_the_middle_(custom_escapes)-8        280µs ± 5%   879µs ± 2%  +214.06%  (p=0.008 n=5+5)
    Scan/all_quotes_(standard_escapes)-8                     332µs ±10%  1106µs ± 1%  +232.77%  (p=0.008 n=5+5)
    Scan/all_quotes_(custom_escapes)-8                       332µs ±10%  1129µs ± 2%  +239.91%  (p=0.008 n=5+5)
    Scan/nothing_but_quotes_(standard_escapes)-8             303µs ± 5%   859µs ± 1%  +183.87%  (p=0.008 n=5+5)
    Scan/nothing_but_quotes_(custom_escapes)-8               303µs ± 5%  1140µs ± 1%  +276.61%  (p=0.008 n=5+5)

This is a pretty significant slowdown -- we've reduced the speed by a
factor of two or three.