batch.Scan: add benchmarks
Since Scan contains the bottleneck for overall parallelization of the
outgoing data, and we're about to add some CPU-intensive work to the
inner loop, add some benchmarks for a variety of use cases.

These are patterned off of (but do not contain any real entries from)
the UK Price Paid Data:

    https://www.gov.uk/guidance/about-the-price-paid-data

The benchmarks also contain two special cases. The first is a warmup
case, to get the garbage collector working before getting numbers.
(Without this case, expect the very first benchmark result to be
artificially slow.) The second special benchmark, in which all columns
consist of nothing but quote characters, provides a worst-case scenario
for SIMD-based optimizations. It's useful for comparisons but should not
probably not be prioritized as an optimization case.