Abort when 'HEADERS' passed to -copy-options
Users familiar with loading CSV data with the COPY command know the
'HEADER' option, and erroneously believe that they should use it with
timescaledb-parallel-copy (by passing 'CSV HEADER' to -copy-options).

timescaledb-parallel-copy handles this use-case with the -skip-header
option. Passing HEADER to -copy-options results in the first row of
every _batch_ being discarded, causing silent data loss.

This change is backwards incompatible, because it aborts on inputs which
it previously accepted.

We have not provided a way to pass HEADER to -copy-options, because we
don't believe that there is a legitimate use-case for this combination.

Fixes: #59