Drop Rewind functionality
After some analysis, the rewind functionality can be easily done given the Location and the source file. Therefore it is not worth the memory overhead created as most of the time the data is not rewind.

a TODO for the future is to have a function that given a Location and the same input, it prints the actual chunk.