feat: add WithQueueSize option to csvcopy.NewCopier
The csvcopy.Copy function sets up one producer and a configurable number
of workers (using WithWorkers). The default number of workers is 8. The
producer transfers batches of CSV rows to be written using a queue. The
default size of the queue is 2*workers.

The WithQueueSize option makes it possible to override the default queue
size.