feat: use pre-allocated buffer pool for batching (#129)
The producer reads up to a pre-configured number of rows or bytes of
data into a buffer, and passes the buffer on to a writer to write to
the database.

In the producer internals, a new buffer would be allocated for each line
which was read from the source CSV. These allocations take time,
increase peak heap usage, and require garbage collection.

This change introduces a pool of buffers which are reused by the
producer. Buffers are created on demand, but allocated to the configured
batchByteSize. The pool has a fixed size of # workers + queue length + 1,
which ensures that the memory usage for the copy operation is bounded at
this batchByteSize * bufferPoolSize.