feat: retry when encountering recoverable errors
When a (potentially) recoverable error is encountered,
timescaledb-parallel-copy will now output that it has encountered an
error and will continue retrying using exponential backoff for up to 5
minutes.

The utility of this change is to bridge over brief database
interruptions, without requiring the copy process to be restarted.

This functionality is on by default, to disable it use the flag
`-disable-retry-on-recoverable-error`.