feat: handle Windows-1252 encoded CSV data for UTF-8 targets
The text in CSV files can be encoded with an arbitrary character set.
When the encoding used in the source file is not compatible with the
target database's encoding, it results either in corrupted data, or an
error is raised by Postgres.

UTF-8 is the most widely used encoding, and we presume that most target
databases today are using it. Windows-1252 is the second most commonly
used encoding.

This commit attempts to provide a fallback to the most common encoding
errors that can occur: it detects when a multibyte encoding error
occurs, and checks if the target is UTF-8. In this case, it configures
Postgres to import the data as though encoded with the Windows-1252
character set.

This behaviour is set to default on, and will only cause overhead when
encoding errors are encountered. If you wish to disable it, use the
flag `-disable-windows-1252-handling`.

Note: this commit purposely does not attempt to determine the "correct"
source encoding and configure Postgres to import the data in that
encoding. The reason is that determining text encoding is based on
heuristics, and tests on small volumes of data did not produce reliable
results.